+++ LED_Matrix/lib/include/PIO/Program/ASM/JMP/JMP.h
@@ -8,18 +8,9 @@
 #define PIO_PROGRAM_ASM_JMP_H
 
 #include "PIO/Program/ASM/ASM.h"
+#include "PIO/Program/ASM/JMP/Conditions.h"
 
 namespace IO {
-    enum class Conditions {
-        NOT_X,
-        X_DEC,
-        NOT_Y,
-        Y_DEC,
-        X_NOT_EQ_Y,
-        PIN,
-        NOT_OSRE
-    };
-
     class JMP : public ASM {
         public:
             JMP(uint8_t addr);
