+++ LED_Matrix/lib/include/PIO/Program/ASM/WAIT/Flags.h
@@ -1,0 +1,18 @@
+/* 
+ * File:   Flags.h
+ * Author: David Thacher
+ * License: GPL 3.0
+ */
+
+#ifndef PIO_PROGRAM_ASM_WAIT_FLAGS_H
+#define PIO_PROGRAM_ASM_WAIT_FLAGS_H
+
+namespace IO {
+    enum class Flags {
+        GPIO,
+        IRQ,
+        PIN
+    };
+}
+
+#endif
